Move component to UI folder
