+++ app/javascript/dashboard/components/ui/Label.vue
