+++ app/javascript/dashboard/components/ui/stories/Label.stories.js
